+++ AttackSurfaceAnalyzer.sln
@@ -19,9 +19,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lib", "Lib\Lib.csproj", "{1A23CFC7-0B8B-44C0-9EB6-5D39BC456CDC}"
-	ProjectSection(ProjectDependencies) = postProject
-		{8D5238C0-803F-42BF-BE0F-479AAD6741E4} = {8D5238C0-803F-42BF-BE0F-479AAD6741E4}
-	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Benchmarks", "Benchmarks\Benchmarks.csproj", "{115B2548-2164-44B6-B0C3-717DB1145550}"
 EndProject
@@ -29,7 +26,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cli", "Cli\Cli.csproj", "{2BE92A57-DAEF-4D62-A5C3-B562F72DC6A2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LogicalAnalyzer", "LogicalAnalyzer\LogicalAnalyzer\LogicalAnalyzer.csproj", "{8D5238C0-803F-42BF-BE0F-479AAD6741E4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorGui", "BlazorGui\BlazorGui.csproj", "{65BA0BAE-57C5-49D4-B952-F743ED0279B4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,10 +50,10 @@ Global
 		{2BE92A57-DAEF-4D62-A5C3-B562F72DC6A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2BE92A57-DAEF-4D62-A5C3-B562F72DC6A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2BE92A57-DAEF-4D62-A5C3-B562F72DC6A2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8D5238C0-803F-42BF-BE0F-479AAD6741E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8D5238C0-803F-42BF-BE0F-479AAD6741E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8D5238C0-803F-42BF-BE0F-479AAD6741E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8D5238C0-803F-42BF-BE0F-479AAD6741E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{65BA0BAE-57C5-49D4-B952-F743ED0279B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{65BA0BAE-57C5-49D4-B952-F743ED0279B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{65BA0BAE-57C5-49D4-B952-F743ED0279B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{65BA0BAE-57C5-49D4-B952-F743ED0279B4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
