+++ Cli/AttackSurfaceAnalyzerClient.cs
@@ -6,7 +6,7 @@ using AttackSurfaceAnalyzer.Utils;
 using CommandLine;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.CST.LogicalAnalyzer;
+using Microsoft.CST.OAT;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
