+++ Lib/Objects/CompareResult.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT License.
 using AttackSurfaceAnalyzer.Types;
-using Microsoft.CST.LogicalAnalyzer;
+using Microsoft.CST.OAT;
 using System.Collections.Generic;
 
 namespace AttackSurfaceAnalyzer.Objects
