+++ Lib/Objects/RuleFile.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT License.
 using AttackSurfaceAnalyzer.Types;
 using AttackSurfaceAnalyzer.Utils;
-using Microsoft.CST.LogicalAnalyzer;
+using Microsoft.CST.OAT;
 using Newtonsoft.Json;
 using Serilog;
 using System;
